Added Comment model and segregated types
