+++ README.md
@@ -1,5 +1,6 @@
 ### This module is wrapper around [JSON placeholder](https://jsonplaceholder.typicode.com/) for testing purposes.
 Current available models: Posts
+
 Posts usage:
 ```ts
 import PostModel from 'jsonplaceholder-wrapper';
