+++ src/constants.ts
@@ -1,5 +1,6 @@
 export const BASE_URL = 'https://jsonplaceholder.typicode.com'
 export const POSTS_URL = `${BASE_URL}/posts`
+export const COMMENTS_URL = `${BASE_URL}/comments`
 export const HEADERS = {
   'Content-type': 'application/json; charset=UTF-8',
 }
