+++ src/models/postModel.ts
@@ -1,12 +1,12 @@
 import { BaseModel } from "./baseModel";
 import { POSTS_URL, HEADERS } from "../constants";
-import { PostType, ResponseType } from "../types";
+import { PostType, ResponseType, PostCreatePaylod, PostUpdatePaylod } from "../types";
 class PostModel extends BaseModel<PostType> {
   constructor() {
     super('posts');
   }
 
-  async create(payload: Omit<PostType, 'id'>): Promise<ResponseType<PostType>> {
+  async create(payload: PostCreatePaylod): Promise<ResponseType<PostType>> {
     const response = await this.fetchData(POSTS_URL, {
       method: 'POST',
       body: JSON.stringify(payload),
@@ -15,7 +15,7 @@ class PostModel extends BaseModel<PostType> {
     return response;
   }
 
-  async update(id: number, payload: Partial<Omit<PostType, 'userId' | 'id'>>): Promise<ResponseType<PostType>> {
+  async update(id: number, payload: PostUpdatePaylod): Promise<ResponseType<PostType>> {
     const response = await this.fetchData(`${POSTS_URL}/${id}`, {
       method: "PUT",
       body: JSON.stringify(payload),
