+++ src/types/album.ts
@@ -1,0 +1,5 @@
+export interface AlbumType {
+  userId: number,
+  id: number,
+  title: string
+}
