+++ src/types/comment.ts
@@ -1,0 +1,11 @@
+export interface CommentType {
+  postId: number,
+  id: number,
+  name: string,
+  email: string,
+  body: string
+}
+
+
+export interface CommentCreatePayload extends Omit<CommentType, 'id'> {}
+export interface CommentUpdatePayload extends Partial<Pick<CommentType, 'name' | 'body'>> {}
