+++ src/types/generics.ts
@@ -1,0 +1,5 @@
+export interface ResponseType<T> {
+  status: number,
+  data?: T[] | T | null,
+  error?: string | Error | unknown
+}
