+++ src/types/photo.ts
@@ -1,0 +1,7 @@
+export interface PhotoType {
+  albumId: number,
+  id: number,
+  title: string,
+  url: string,
+  thumbnailUrl: string
+}
