+++ src/types/todo.ts
@@ -1,0 +1,6 @@
+export interface TodoType {
+  userId: number,
+  id: number,
+  title: string,
+  completed: boolean
+}
